Spark_kafka_streaming/requirements.txt
kafka-python==2.0.2
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
pyspark==3.0.3